Fixed evaluation of nested conditionals
We need to change the order in which the conditionals are evaluated to prevent
that already disabled fields get reenabled again. This happens in nested
conditionals. Consider C1 is nesting C2 and C1 is True and C2 is False. Then
when evaluating C2 first the fields withing C2 gets reenabled by C1 again. If
C1 is evaluated first the conditionals work as expected.