Simplify formbar/config.py(handle_entity_prefix)
The reasoning in this function is too complex.
Simply use the re.sub() function instead of the
sorting and the iterated use of the str.replace() function.

Pull the compilation of the regular expression out
of the inner loop to a global as it never changes.

Allow '-' in variable names like brabbel does.

Only run once over the tree to avoid three expensive
findall calls.