Make deserialisation on field init more robust
Deserialisation is not yet supported for relations as the relation names are
missing at this point. To fix this we need to recator the code to much. This is
not worth the effort as currently we do not have any usecase where is is
needed, and this should be fixed in the pending refactoring in refactor/fields
branch already.