Fixed layout error for indented form fields.
Replaced use of margin with padding. This fixes some sizing issues of fields
in case the label and field was indeted. Beside fields breaking the layout by
breaking the max page width there was an error for fields which has been
aligned side by side. In this case the first fields gets overlapped by the
second field. In case of select fields the dropdown was not usable anymore
because it was overlappend by the label of the next item.